tuple: Adds Quad and Quin
As the final (I hope) commit for tuples, we add Quad and Quin, placing a
style pattern to at least reduce the thinking required to implement
Comparable, in addition to adding factory methods in the Tuples class.

Trio is also renamed to Triple, formatting is updated in the Tuples
class (break after generic signature) and methods are renamed to be a
single overload.

It is not intended to implement further type-safe tuple classes, as
beyond more than 5 items, custom objects should probably be used, however
the Tuple base class is designed for these scenarios and should exhibit
correct behaviour with respect to higher order functions.