tuples: Protect invariants
It occurred to me that the Tuple class was doing nothing to prevent its
Items from being modified. In this commit, we wrap the incoming array of
elements in an unmodifiable list to ensure the structure is read-only.
Unfortunately this loses the terse access syntax, though the Items
collection is not intended for common public use.

Finally, we add a static test routine for some basic properties of
OrderedTriple, this method is a pre-cursor to a unit test.